Changed: renamed `DomConverter.isCorrectDomSelection(Position)` to `DomConverter.isDomSelection(Position)Correct`. Changed: moved `jumpOverUiElement` handler to `uielement.js`. Docs: minor fixes.
